+++ lib/bind3.dart
@@ -1,9 +1,9 @@
+// ignore: unnecessary_library_name
 library bind3;
 
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
-import 'package:auto_route/auto_route.dart' as i1;
 
 part 'app_router_base.dart';
 part 'route_core.dart';