+++ lib/bind3.gr.dart
@@ -1,3 +1,5 @@
+part of 'bind3.dart';
+// dart format width=80
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
 // **************************************************************************
@@ -6,13 +8,3 @@
 
 // ignore_for_file: type=lint
 // coverage:ignore-file
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-part of "bind3.dart";
-
-abstract class $EmptyRouter extends i1.RootStackRouter {
-  $EmptyRouter({super.navigatorKey});
-
-  @override
-  final Map<String, i1.PageFactory> pagesMap = {};
-}